refactor: input 컴포넌트 내부 로직 커스텀 훅으로 분리
